Refactor: 컴포넌트 디렉토리에 common 디렉토리 생성
- common 디렉토리에 atom 단위 컴포넌트를 이동시켰습니다.
