+++ client/src/components/main/CreateForm.tsx
@@ -4,7 +4,7 @@ import styled, { css } from 'styled-components';
 import useInput from '@hooks/useInput';
 import { postRoom } from '@utils/apis';
 import { useUser } from '@contexts/userContext';
-import Select from '@components/Select';
+import Select from '@components/common/Select';
 import { adminOptions } from '@utils/utils';
 
 const Container = styled.div`
