+++ client/src/components/main/SideBar.tsx
@@ -3,7 +3,7 @@ import styled, { css } from 'styled-components';
 import LoginModal from '../LoginModal';
 import { useUser, useUserFns } from '@contexts/userContext';
 import { IoLogOutOutline } from 'react-icons/io5';
-import Modal from '@components/Modal';
+import Modal from '@components/common/Modal';
 import CreateForm from './CreateForm';
 import { setCookie } from '@utils/cookie';
